Update quantities for moves to create when computing internal shipments
Since changeset 2debb3c82659, there is a loop over the lead time. But as the
shipments are created after the loop, the computation of quantities inside the
loop does not include the created moves.
Instead we compute only the delta for each iteration and update the quantities
also with the created moves.

issue6152
review27901002